Update StringBasics operator= to use a long long instead of unsigned long

+++ general/StringBasics.cpp
@@ -289,7 +289,7 @@ String & String::operator = (unsigned int rhs)
 {
     Clear();
 
-    unsigned long  base = 10;
+    unsigned long long base = 10;
     int   digits = 1;
 
     while (rhs >= base)
